(Temporary) only compute S_AF for f-electrons in PAM

+++ src/dqmc_struct.F90
@@ -92,8 +92,8 @@ module DQMC_STRUCT
      integer, pointer  :: chi_phase(:,:) 
                               
      ! decide (-1)^x+y for computing S_AF and S_CDW in plane
-     ! for bilayer cases, only include z=0 component
-     integer, pointer  :: AFphase(:)
+     ! Defined in dqmc_geom_wrap.F90
+     real(wp), pointer  :: AFphase(:)
 
      integer           :: n_b           ! number of neighbors types
      type(CCS)         :: B             ! Neighborhood matrix
